add extension to move big chunks of code from program.cs
